Picture, description, ingredients, steps
Picture of strawberry bruschetta for the user to have a visual description.
Image used from img directory.
Description of strawberry bruschetta for the user to get a more detailed
description. Two paragraph elements used to create descriptions.
Ingredients list for the user to aquire needed items. Unordered list
used to create list.
Steps list for the user to follow in making bruschetta. Ordered list for
user to follow in steps.